+++ app/reaction/suggestion-box.js
@@ -1,5 +1,5 @@
 const BOT_ROLE_NAME = "BOT";
-const common = require("./common.js");
+const common = require("../common.js");
 const suggestionChannelParentId = process.env.CATEGORY_PARENT_ID_SUGGESTION_BOX;
 const infomationChannelId = process.env.CHANNEL_ID_INFOMATION;
 const Discord = require("discord.js");
